+++ webpack.config.js
@@ -6,6 +6,27 @@ module.exports = {
     popup: './src/popup/popup.js',
     "bandcamp-content": './src/bandcamp/modules/main.js'
   },
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: {
+          loader: 'ts-loader',
+          options: {
+            transpileOnly: true
+          }
+        },
+        exclude: /node_modules/,
+      },
+      {
+        test: /\.js$/,
+        loader: "source-map-loader"
+      },
+    ],
+  },
+  resolve: {
+    extensions: ["", ".webpack.js", ".web.js", ".ts", ".tsx", ".js"],
+  },
   output: {
     filename: '[name].js',
     path: path.resolve(__dirname, 'dist'),